refactor: move utilities/domain and utilities/url to packages
