+++ apps/web/components/Pages/SignUp/index.tsx
@@ -12,7 +12,7 @@ import TextField from 'components/TextField';
 import EmailField from 'components/EmailField';
 import { Button } from '@linen/ui';
 import PasswordField from 'components/PasswordField';
-import { qs } from 'utilities/url';
+import { qs } from '@linen/utilities/url';
 import GitHubButton from 'components/Auth/GitHubButton';
 
 export interface SignUpProps {
