+++ apps/web/pages/api/forgot-password/index.ts
@@ -2,7 +2,7 @@ import { NextApiRequest, NextApiResponse } from 'next/types';
 import ResetPasswordMailer from 'mailers/ResetPasswordMailer';
 import { generateToken } from 'utilities/token';
 import { prisma } from '@linen/database';
-import { getCurrentUrl } from 'utilities/domain';
+import { getCurrentUrl } from '@linen/utilities/domain';
 
 async function create(request: NextApiRequest, response: NextApiResponse) {
   const { email } = JSON.parse(request.body);
